Extract `Authentication`
