+++ src/ZoomHub/API.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -12,7 +11,6 @@ where
 
 import Control.Monad.Except (throwError, void)
 import Control.Monad.IO.Class (liftIO)
-import Data.Aeson (FromJSON, ToJSON)
 import qualified Data.ByteString.Char8 as BC
 import Data.Foldable (fold, for_)
 import Data.HashMap.Strict (HashMap)
@@ -24,7 +22,6 @@ import qualified Data.Text as T
 import Data.Text.Encoding (encodeUtf8)
 import qualified Data.Time as Time
 import qualified Data.UUID.V4 as UUIDV4
-import GHC.Generics (Generic)
 import Network.Minio as Minio
   ( awsCI,
     fromAWSEnv,
@@ -46,8 +43,7 @@ import qualified Network.URI.Encode as URIEncode
 import Network.Wai (Application)
 import Network.Wai.Middleware.Cors (simpleCors)
 import Servant
-  ( BasicAuthData (BasicAuthData),
-    Capture,
+  ( Capture,
     Context (EmptyContext, (:.)),
     Get,
     Handler,
@@ -68,13 +64,8 @@ import Servant.Auth.Server
   ( Auth,
     AuthResult (Authenticated, BadPassword, Indefinite, NoSuchUser),
     BasicAuth,
-    BasicAuthCfg,
-    FromBasicAuthData,
-    FromJWT,
-    ToJWT,
     defaultCookieSettings,
     defaultJWTSettings,
-    fromBasicAuthData,
     generateKey,
   )
 import Servant.HTML.Lucid (HTML)
@@ -99,6 +90,7 @@ import ZoomHub.API.Types.NonRESTfulResponse
     mkNonRESTful404,
     mkNonRESTful503,
   )
+import qualified ZoomHub.Authentication as Authentication
 import ZoomHub.Config (Config)
 import qualified ZoomHub.Config as Config
 import qualified ZoomHub.Config.AWS as AWS
@@ -119,8 +111,6 @@ import ZoomHub.Storage.PostgreSQL as PG
     markAsSuccess,
     markAsVerified,
   )
-import ZoomHub.Types.APIUser (APIUser)
-import qualified ZoomHub.Types.APIUser as APIUser
 import ZoomHub.Types.BaseURI (BaseURI, unBaseURI)
 import qualified ZoomHub.Types.Content as Internal
 import ZoomHub.Types.ContentBaseURI (ContentBaseURI)
@@ -139,17 +129,6 @@ import ZoomHub.Web.Static (serveDirectory)
 import ZoomHub.Web.Types.Embed (Embed, mkEmbed)
 import ZoomHub.Web.Types.EmbedDimension (EmbedDimension)
 import ZoomHub.Web.Types.EmbedId (EmbedId, unEmbedId)
-
-data AuthenticatedUser = AuthenticatedUser
-  deriving (Show, Generic)
-
-instance ToJSON AuthenticatedUser
-
-instance FromJSON AuthenticatedUser
-
-instance ToJWT AuthenticatedUser
-
-instance FromJWT AuthenticatedUser
 
 -- API
 type API =
@@ -207,7 +186,7 @@ type API =
       :> Capture "token" String
       :> Put '[JSON] Content
     -- API: RESTful: Completion
-    :<|> Auth '[BasicAuth] AuthenticatedUser
+    :<|> Auth '[BasicAuth] Authentication.AuthenticatedUser
       :> "v1"
       :> "content"
       :> Capture "id" ContentId
@@ -294,20 +273,6 @@ server config =
     uploads = Config.uploads config
     viewerScript = Config.openSeadragonScript config
 
-type instance BasicAuthCfg = BasicAuthData -> IO (AuthResult AuthenticatedUser)
-
-authCheck :: APIUser -> BasicAuthData -> IO (AuthResult AuthenticatedUser)
-authCheck apiUser (BasicAuthData unverifiedUsername unverifiedPassword) =
-  let username = APIUser.username apiUser
-      password = APIUser.password apiUser
-   in if username == lenientDecodeUtf8 unverifiedUsername
-        && password == lenientDecodeUtf8 unverifiedPassword
-        then pure $ Authenticated AuthenticatedUser
-        else pure NoSuchUser
-
-instance FromBasicAuthData AuthenticatedUser where
-  fromBasicAuthData authData authCheckFunction = authCheckFunction authData
-
 -- App
 app :: Config -> IO Application
 app config = do
@@ -316,7 +281,10 @@ app config = do
   where
     -- TODO: Can we use `BasicAuth` without JWT and cookies?
     cfg jwtKey =
-      defaultJWTSettings jwtKey :. defaultCookieSettings :. authCheck (Config.apiUser config) :. EmptyContext
+      defaultJWTSettings jwtKey
+        :. defaultCookieSettings
+        :. Authentication.check (Config.apiUser config)
+        :. EmptyContext
     logger = Config.logger config
 
 -- Handlers
@@ -401,7 +369,7 @@ restUpload baseURI awsConfig uploads email =
       noNewContentErrorAPI
     UploadsEnabled -> do
       currentTime <- liftIO Time.getCurrentTime
-      s3UploadKey <- (T.pack . show) <$> liftIO UUIDV4.nextRandom
+      s3UploadKey <- T.pack . show <$> liftIO UUIDV4.nextRandom
       let expiryTime = Time.addUTCTime Time.nominalDay currentTime
           key = "uploads/" <> s3UploadKey
           s3BucketURL = "https://" <> s3Bucket <> ".s3.us-east-2.amazonaws.com"
@@ -478,7 +446,7 @@ restContentCompletionById ::
   BaseURI ->
   ContentBaseURI ->
   Pool Connection ->
-  AuthResult AuthenticatedUser ->
+  AuthResult Authentication.AuthenticatedUser ->
   ContentId ->
   ContentCompletion ->
   Handler Content
